maybe he wants to be a float. /google..

+++ task_03.py
@@ -71,8 +71,8 @@ REPORT = ('Loan report for: {0} \n'
 LOAN = '{:0,d}'.format(LOAN)
 TOTAL = '{:0,d}'.format(TOTAL)
 
-LOAN = int(LOAN)
-TOTAL = int(TOTAL)
-YEARS = int(YEARS)
+LOAN = float(LOAN)
+TOTAL = float(TOTAL)
+YEARS = float(YEARS)
 
 print REPORT.format(NAME, LOAN, YEARS, QUALIFIED, TOTAL)
